requirements.txt
streamlit==1.8.1
numpy==1.21.5
pandas==1.4.1
plotly==5.6.0
opencv-python-headless==4.5.5.64
Pillow==9.0.1